+++ src/nestjs/factory/app-factory.spec.ts
@@ -1,18 +1,21 @@
 import { Controller, Get, INestApplication, Module } from '@nestjs/common';
 import { ConfigService } from '@nestjs/config';
-import request from 'supertest';
+import { PinoLogger } from 'nestjs-pino';
+import supertest from 'supertest';
+import { getLoggedObjects, spyOnLogger } from '../../logging/testing.js';
 import { createApp } from './app-factory.js';
 
 @Controller('test')
 class TestController {
   readonly someConfValue: string;
 
-  constructor(configService: ConfigService) {
+  constructor(readonly logger: PinoLogger, configService: ConfigService) {
     this.someConfValue = configService.getOrThrow('SOME_CONF_VALUE');
   }
 
   @Get('/')
   async get() {
+    this.logger.info('💮');
     return;
   }
 }
@@ -28,11 +31,14 @@ class AppModule {}
 describe('app-factory', () => {
   let app: INestApplication;
   let previousEnv: NodeJS.ProcessEnv;
+  let request: supertest.SuperTest<supertest.Test>;
 
   beforeEach(async () => {
+    spyOnLogger();
     previousEnv = { ...process.env };
     process.env.SOME_CONF_VALUE = '🔧';
     app = await createApp(AppModule);
+    request = supertest(app.getHttpServer());
   });
 
   afterEach(async () => {
@@ -41,9 +47,7 @@ describe('app-factory', () => {
   });
 
   it('should not return the x-powered-by header', async () => {
-    const actualResponse = await request(app.getHttpServer())
-      .get('/test')
-      .expect(200);
+    const actualResponse = await request.get('/test').expect(200);
     expect(actualResponse.headers).not.toHaveProperty('x-powered-by');
   });
 
@@ -59,5 +63,23 @@ describe('app-factory', () => {
     const actualConfValue = controller.someConfValue;
 
     expect(actualConfValue).toEqual('🔧');
+  });
+
+  it('should expose the logger module', async () => {
+    const controller = app.get(TestController);
+
+    const actualLogger = controller.logger;
+    await request.get('/test').expect(200);
+
+    expect(actualLogger).toBeInstanceOf(PinoLogger);
+    expect(getLoggedObjects({ predicate: (o) => o.message === '💮' })).toEqual([
+      expect.objectContaining({
+        req: expect.objectContaining({ url: '/test' }),
+      }),
+    ]);
+  });
+
+  it('should import the healthcheck module', async () => {
+    await request.get('/health').expect(200);
   });
 });