+++ src/nestjs/healthcheck/index.ts
@@ -1,0 +1,1 @@
+export { HealthcheckModule } from './healthcheck.module.js';