+++ CHANGELOG.md
@@ -21,3 +21,4 @@ Features:
 - Implement the `LockManager`.
 - Implement the NestJS application factory and corresponding testing utilities.
 - Implement the `LoggerModule`.
+- Implement the `HealthcheckModule`.